% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_available.R
\name{trans_available}
\alias{trans_available}
\title{Display available locales}
\usage{
trans_available()
}
\value{
Returns available locales in df
}
\description{
This function displays available locales
}
\examples{
path <- system.file("examples", package = "translated")
trans_path(path)

# Display available locales
trans_available()

}
