% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_available_locales.R
\name{trans_available_locales}
\alias{trans_available_locales}
\title{Display available locales}
\usage{
trans_available_locales()
}
\value{
Returns available locales in df
}
\description{
This function displays available locales
}
\examples{
library(translated)
path <- system.file("examples", package = "translated")
trans_path(path)

# Display available locales
trans_available_locales()

}
