% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_path.R
\name{trans_path}
\alias{trans_path}
\title{Access translation path}
\usage{
trans_path(path)
}
\arguments{
\item{path}{\code{character(1)}\cr
Path to set, should be a folder containing JSON files (possibly in
subdirectories).}
}
\value{
If \code{path} was not passed, currently set path, else nothing.
}
\description{
This function allows setting translation path and accessing the
currently set one.
}
\examples{
# Set path to example translations shipped with this package
trans_path(system.file("examples", package = "translated"))

# Check the current path
trans_path()

}
