% Generated by roxygen2: do not edit by hand
\name{translate_me}
\alias{translate_me-es}
\title{Título de la documentación}
\description{
Título un poco más largo
}
\details{
@param x un parámetro
 @param b otro parámetro

 @details
 Muchos detalles! Con **markdown**!
 
 @section Sección especial:

 Esta es una sección especial

 @examples

 translate_me()  
}
