% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate-me.R
\name{translate_me}
\alias{translate_me}
\title{Title of documentation}
\usage{
translate_me(x = 1, b = 2)
}
\description{
Slightly longer description
}
\details{
@param x a parameter
 @param b more parameter

 @details
 Details ahoy! With **markdown**!

 @section Custom section:

 This is a custom section.

 @examples

 translate_me()
}
