% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_reload.R
\name{trans_reload}
\alias{trans_reload}
\title{Force translation reload}
\usage{
trans_reload()
}
\value{
No return value, called for its side effect.
}
\description{
This function forces the translations to be loaded from JSON
files again.
}
\examples{
trans_path(system.file("examples", package = "translated"))
trans_reload()

}
